requirements.txt
flask
Flask-SQLAlchemy
Werkzeug
pyjwt
flask_restx
psycopg2